Add getenv support to the Java payload
This will bring this meterpreter in line with the other meterpreter
payloads so that the getenv feature works consistently across them.

I also had to tweak the way the TLVs were stored so that we could
support the notion of having more than one value appear that is of
the same type.