Transport refactoring, dispatch functionality, and more
This commit adds a new idea to the dispatch loop that allows commands to
exit. It also adds the core_shutdown command so that it functions
correct. There are a bunch of other changes around transports as well,
and this commit adds a "resilient" TCP transport.

HTTP/S to come. progress being made. This commit is messy with debug
statements all through it, and they'll get removed down the track.